Add mif2json.py
mif2json.py is the begginings of a script to convert
Maker (FrameMaker) Interchange Format (MIF) files to JSON.

Currently, it is heavily dependent on the source of the Power ISA document
and only extracts instructions and some related data.

It is a work in progress.

Also adding an associated LICENSE file (Apache v2).

Signed-off-by:  Paul A. Clarke <pc@us.ibm.com>