Add mnemonics, pseudocode, and textual description to output
Each instruction when selected, will display the instruction mnemonic(s),
registers, the pseudocode explaining what the instruction does, and the
textual description of the instruction.

The JSON for the "registers" was badly formatted, and now fixed.

Also, some bad data (where the instruction mnemonic and registers were
in a "Body" paragraph instead of an "Instruction Form" paragraph) has
been mildly accommodated so that a "regs" field is always present,
although it will be empty.

I have asked that the canonical data source be fixed.

Actual display formatting is still a work in progress.  :-/

Signed-off-by:  Paul A. Clarke <pc@us.ibm.com>