Add filter by ISA level
Adding checkbox for each ISA level, so that instructions can be filtered.

ISA levels are hardcoded instead of derived from ISA.json, because it is
impossible to know the order otherwise.  (Maybe they could be pulled from
a MIF file and put in their own JSON file?)

An "all" checkbox is also added to facilitate quickly selecting or
deselecting all other checkboxes.

Signed-off-by:  Paul A. Clarke <pc@us.ibm.com>