Fix instruction layouts with extra fields
Some instructions were pulling in extra fields because the `TblID`
was present in multiple chapters, but were being replicated because
the `ISA.json` file was being created by concatenating the file
...because `mif2json.py` did not support processing multiple files
simultaneously:
```
./utils/mif2json.py <(Chapter1.mif Chapter2.mif Appendix.mif)
```

Fix this by adding support for processing multiple files if provided
on a single command line:
```
./utils/mif2json.py Chapter1.mif Chapter2.mif Appendix.mif
```

The above fix introduced a new problem where subsequent chapter
introductory text was included in the body of the last instruction
from the previous chapter, because it's all just `Para` content.

Fix that by trying to track whether the current state of parsing
is likely within an instruction, and resetting that to `False`
and the end of each chapter.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>