HTML-escape JSON "code" and "body" content
Subscripts and superscripts do not work within Carbon `CodeSnippet`
components, unless you use `dangerouslySetInnerHTML` properties.
As the name of the property obviously implies, this is not always
safe, and the HTML needs to be clean.  Since the content in
question is produced by `utils/mif2json.py`, make sure to HTML
escape it during the translation.

It may be worth considering the use of an HTML sanitizer if the
content gets much more sophisticated.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>