Add Power ISA level to each mnemonic
On the main display, add the level(s) of the Power ISA in which the
mnemonic(s) were introduced.
```
Description **mnemonic operand,operand** ISA-level
```

Although this does not occur with the current level of the ISA,
an allowance is made for when a single instruction description has
multiple mnemonics and some mnemonics differ from others as to
when they were introduced.  The set of _unique_ ISA levels is
displayed as a comma-separated list.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>