Use CodeSnippet "inline" for better information density
Using `<CodeSnippet type="multi">` had a few problems.
- Clicking on the "copy" button would copy _all_ mnemonics, when that
would never be wanted (ignoring the fact that the "copy" button does
actually work without more wiring).
- There is lots of extra space, both to the right and below, around
the text.

Using `<CodeSnippet type="inline">` is much more compact, and allows
for easy embedding in a table, such that the mnemonics selectors /
conditions can be moved outside of the text box, and the "copy" button
(when it is made to work) will just copy the {mnemonic + operands},
which is reasonably close to a useful string.

Bolding added, and the selectors / conditions text size is reduced to
make it not dominate the line of text.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>