Enable copy to clipboard
The mnemonics and pseudocode displays are in `CodeSnippet` components
and the "copy to clipboard" function is not automatically wired up.
See https://github.com/carbon-design-system/carbon/issues/6995.

Referenced in that issue is one solution, a component called
`react-copy-to-clipboard`.  With help, this is almost trivial to
implement (the documentation and examples are almost non-existant).

- Utilize `CopyToClipboard` components for the two aforementioned
`CodeSnippet` components.

- Add some `key` attributes to silence some JavaScript warnings.

- Add `tbody` markup to silence a JavaScript warning.

- Remove spaces between instruction opcodes to make copy-paste
more useful, as assembly code does not generally have spaces there.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>