Make associated entities into live links
For each associated intrinsic or instruction, turn the displayed
mnemonic into a live link that, when clicked navigates to that
entity.  Well, almost:  the link enables searching for the entity
type using the string of the entity mnemonic.  If there are multiple
hits for that criteria, all are listed as they would be with a
manual search.

Known issue: the checkboxes and search field as displayed are not
updated with the manufactured criteria. (This could be considered
a feature, as it might be nice to know the starting point.)

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>