Add intrinsic type signatures
Include a new Carbon component, DataTable, to display the type signatures
for each intrinsic.

The data for a DataTable component requires a list of header objects, each
with a "key" and "header" fields.  In addition, each row in the table is
an object with a unique "id" field, and fields corresponding the each "key"
in the header.

So, add a new script, `utils/addVIPRids.py` to reformat the ISA.json after
the merge.

One bug was fixed in the parser for the Power Vector Intrinsics Programming
Reference, an independent project, the result of which is incorporated here.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>