Fix `sh` fields from `SH` operand
`SH` operand gets split into two fields: the high-order bit
and the remaining 5 bits.  Both of these fields are named
`sh`, which makes parsing challenging.  For example, which bit of `SH`
is the single-bit `sh`?  It's impossible to know without apriori
knowledge, which obviates the whole point of parsing.

Rename the two `sh` fields to be `sh` (as in "shift-high") for the high
order bit, and `sl` (as in "shift-low") for the remaining bits.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>