Clarify `target_addr` operands
There are two instructions which accept a `target_addr` operand,
but their semantics are signficantly different:

- "Branch" accepts a 26-bit `target_addr` operand, that is
  truncated (by shifting right 2 bits) to 24 bits and stored in
  the `LI` field.  Upon use, the field is shifted left 2 bits
  and sign-extended to 64 bits.

- "Branch Conditional" accepts a 16-bit `target_addr` operand, that
  is truncated (by shifting right 2 bits) to 14 bits and stored in
  the `BD` field.  Upon use, the field is shifted left 2 bits
  and sign-extended to 64 bits.

Given the differences, it is helpful when parsing if these two
operands had different names.

Change the operands to `target_addr26` and `target_addr16`, respectively.

Changes here are the result of parsing updated ISA source.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>