Add `value` to Simple Branch Mnemonic operands
A means is needed to disassemble an opcode to an extended mnemonic.
This includes operands, of course.  There is no intuitive way to
translate a Condition Register bit number to a Condition Register
field number, especially in the absence of any knowledge about
what those fields and operands represent.

Add information to describe the direct transformation from an
opcode to operands.

Changes reflect regenerating `emnem.json`.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>