Add new LaTeX2json.py
The OpenPOWER Foundation became the owner of the Power ISA.

With that change in maintainership came a change in the source document
format for the ISA.  This project is based on the original Framemaker-
based source (converted to a more parsable markup format called Maker
Interchange Format, or MIF). The OpenPOWER Foundation ISA Technical
Working group changed from the proprietary Framemaker/MIF format to a
well- established text-formatting (de facto) standard, LaTeX. LaTeX is,
however, not a database representation, per se. The author of this
project has been working within the OpenPower Foundation ISA Technical
Working Group to add LaTeX-based "markup" to facilitate transformation
of salient parts of the Power ISA textual documents to a JSON database
for the Power ISA instructions.

This is the Python-based parser for the new LaTeX-based Power ISA,
transforming key contents from LaTeX to JSON.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>