Fix Books CheckBox
The `key` and `id` for the CheckBox for each book is using the shortname,
which can be as short as one character. This seems to cause problems, likely
conflicting with some other object with a stupidly short name.

Prefix the `key` and `id` for the CheckBox for each book with "Book".

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>